moved feature flags to demo app
