+++ azure-communication-ui/azure-communication-ui-demo-app/src/calling/java/com/azure/android/communication/ui/callingcompositedemoapp/launcher/FeatureFlagView.kt
@@ -7,7 +7,7 @@ import android.content.Context
 import android.util.AttributeSet
 import android.widget.CheckBox
 import android.widget.LinearLayout
-import com.azure.android.communication.ui.calling.utilities.implementation.FeatureFlags
+import com.azure.android.communication.ui.callingcompositedemoapp.features.FeatureFlags
 
 // This lists all the Features in the FeatureFlag system
 // and lets you enable/disable them.
